Rename style.js to styles.js
