+++ src/components/Home/styles.js
